chore: move ProjectTargetSelector into a separate file
